apps/vis_art/requirements.txt
streamlit
pandas
numpy
pillow
tensorflow
